Convert DeckGlMap component to typescript Closes #769
